refactor: separate View into InputView OutputView
