+++ src/main/kotlin/lotto/view/View.kt
@@ -4,70 +4,32 @@ import lotto.domain.Lotto
 import lotto.domain.LottoResult
 import lotto.domain.Lottos
 import lotto.domain.WinningLotto
-import kotlin.math.floor
 
 object View {
     fun readPrice(): Int {
-        println(MESSAGE_ENTER_PRICE)
-        val input: String = readLine() ?: throw IllegalArgumentException(ERROR_MESSAGE_INVALID_INPUT_STATE)
-        val price: Int = input.toIntOrNull() ?: throw IllegalArgumentException(ERROR_MESSAGE_INPUT_NOT_A_NUMBER)
-        return price
+        OutputView.requestPrice()
+        val price: Int = InputView.readPrice()
         val lottoCount: Int = price / Lotto.LOTTO_PRICE
-        println("${lottoCount}개를 구매했습니다.")
+        OutputView.showLottoCount(lottoCount)
+        return price
     }
 
     fun showLottos(lottos: Lottos) {
-        lottos.lottos.forEach { lotto: Lotto ->
-            println(lotto.numbers.sorted())
-        }
-    }
-
-    fun readWinningLotto(): WinningLotto {
-        println(MESSAGE_ENTER_WINNING_NUMBERS)
-        val lottoInput: String = readLine() ?: throw IllegalArgumentException(ERROR_MESSAGE_INVALID_INPUT_STATE)
-        val wonLotto =
-            Lotto(
-                lottoInput
-                    .split(LOTTO_NUMBERS_DELIMITER)
-                    .map {
-                        it.trim().toIntOrNull() ?: throw IllegalArgumentException(MESSAGE_INVALID_WINNING_NUMBERS)
-                    }.toSet(),
-            )
-        println(MESSAGE_ENTER_BONUS_NUMBER)
-        val bonusInput: String = readLine() ?: throw IllegalArgumentException(ERROR_MESSAGE_INVALID_INPUT_STATE)
-        val bonusNumber: Int =
-            bonusInput.toIntOrNull() ?: throw IllegalArgumentException(ERROR_MESSAGE_INPUT_NOT_A_NUMBER)
-        return WinningLotto(wonLotto, bonusNumber)
+        OutputView.showLottos(lottos)
     }
 
     fun showResult(
         lottoResults: List<LottoResult>,
         profitRate: Double,
     ) {
-        println(MESSAGE_RESULT_HEADER)
-        println("3개 일치 (5000원) - ${lottoResults.count { lottoResult: LottoResult -> lottoResult == LottoResult.FIFTH }}개")
-        println("4개 일치 (50000원) - ${lottoResults.count { lottoResult: LottoResult -> lottoResult == LottoResult.FOURTH }}개")
-        println("5개 일치 (1500000원) - ${lottoResults.count { lottoResult: LottoResult -> lottoResult == LottoResult.THIRD }}개")
-        println("5개 일치, 보너스 볼 일치(30000000원) - ${lottoResults.count { lottoResult: LottoResult -> lottoResult == LottoResult.SECOND }}개")
-        println("6개 일치 (2000000000원) - ${lottoResults.count { lottoResult: LottoResult -> lottoResult == LottoResult.FIRST }}개")
-        println("총 수익률은 ${floor(profitRate * 100) / 100}입니다.(기준이 1이기 때문에 결과적으로 ${makeProfitRateDescription(profitRate)} 의미임)")
+        OutputView.showResult(lottoResults, profitRate)
     }
 
-    private fun makeProfitRateDescription(profitRate: Double): String =
-        when {
-            profitRate < 1.0 -> "손해라는"
-            profitRate == 1.0 -> "본전이라는"
-            profitRate > 1.0 -> "이득이라는"
-            else -> throw IllegalStateException(ERROR_MESSAGE_ILLEGAL_STATE)
-        }
-
-    private const val LOTTO_NUMBERS_DELIMITER = ","
-    private const val MESSAGE_ENTER_WINNING_NUMBERS = "지난 주 당첨 번호를 입력해 주세요."
-    private const val MESSAGE_ENTER_BONUS_NUMBER = "보너스 볼을 입력해 주세요."
-    private const val MESSAGE_ENTER_PRICE = "구입금액을 입력해 주세요."
-    private const val MESSAGE_INVALID_WINNING_NUMBERS = "양식에 맞게 입력해주세요. (1, 2, 3, 4, 5, 6)"
-    private const val MESSAGE_RESULT_HEADER = "당첨 통계\n---------"
-    private const val ERROR_MESSAGE_INVALID_INPUT_STATE = "정상적으로 입력되지 않았습니다."
-    private const val ERROR_MESSAGE_INPUT_NOT_A_NUMBER = "숫자를 입력해주세요."
-    private const val ERROR_MESSAGE_ILLEGAL_STATE = "연산에 실패했습니다."
+    fun readWinningLotto(): WinningLotto {
+        OutputView.requestWinningLotto()
+        val winningLotto = InputView.readWinningNumbers()
+        OutputView.requestBonusNumber()
+        val bonusNUmber = InputView.readBonusNumber()
+        return WinningLotto(winningLotto, bonusNUmber)
+    }
 }
